requirements.txt
# Playwright
playwright==1.40.0

# Pandas e processamento de dados
pandas==2.1.4
polars==0.20.3
openpyxl==3.1.2  # Para ler Excel

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Utils
requests==2.31.0
pydantic==2.5.3
